test(cdk/drag-drop): move transform layout tests out of the common ones
Moves the tests that use `transform` to lay out the list out of the common ones since they're specific to the `SingleAxisSortStrategy`.
