+++ src/cdk/drag-drop/directives/drop-list-shared.spec.ts
@@ -47,16 +47,17 @@ import {
   stopDraggingViaTouch,
   tickAnimationFrames,
 } from './test-utils.spec';
+import {NgFor} from '@angular/common';
 
-const ITEM_HEIGHT = 25;
-const ITEM_WIDTH = 75;
+export const ITEM_HEIGHT = 25;
+export const ITEM_WIDTH = 75;
 
 export function defineCommonDropListTests(config: {
   /** Orientation value that will be passed to tests checking vertical orientation. */
-  verticalListOrientation: DropListOrientation;
+  verticalListOrientation: Exclude<DropListOrientation, 'horizontal'>;
 
   /** Orientation value that will be passed to tests checking horizontal orientation. */
-  horizontalListOrientation: DropListOrientation;
+  horizontalListOrientation: Exclude<DropListOrientation, 'vertical'>;
 
   /** Asserts that sorting an element up works correctly. */
   assertUpwardSorting: (fixture: ComponentFixture<unknown>, items: Element[]) => void;
@@ -70,6 +71,12 @@ export function defineCommonDropListTests(config: {
   /** Gets the siblings of an element, sorted by their visible position. */
   getElementSibligsByPosition: (element: Element, direction: 'top' | 'left') => Element[];
 }) {
+  const {
+    DraggableInHorizontalDropZone,
+    DraggableInScrollableHorizontalDropZone,
+    DraggableInHorizontalFlexDropZoneWithMatchSizePreview,
+  } = getHorizontalFixtures(config.horizontalListOrientation);
+
   function createComponent<T>(
     type: Type<T>,
     testConfig: DragDropTestConfig = {},
@@ -1623,201 +1630,26 @@ export function defineCommonDropListTests(config: {
       }),
     );
 
-    it('should lay out the elements correctly, when swapping down with a taller element', fakeAsync(() => {
-      const fixture = createComponent(DraggableInDropZone);
-      fixture.detectChanges();
-
-      const items = fixture.componentInstance.dragItems.map(i => i.element.nativeElement);
-      const {top, left} = items[0].getBoundingClientRect();
-
-      fixture.componentInstance.items[0].height = ITEM_HEIGHT * 2;
-      fixture.changeDetectorRef.markForCheck();
-      fixture.detectChanges();
-
-      startDraggingViaMouse(fixture, items[0], left, top);
-
-      const placeholder = document.querySelector('.cdk-drag-placeholder')! as HTMLElement;
-      const target = items[1];
-      const targetRect = target.getBoundingClientRect();
-
-      // Add a few pixels to the top offset so we get some overlap.
-      dispatchMouseEvent(document, 'mousemove', targetRect.left, targetRect.top + 5);
-      fixture.detectChanges();
-
-      expect(placeholder.style.transform).toBe(`translate3d(0px, ${ITEM_HEIGHT}px, 0px)`);
-      expect(target.style.transform).toBe(`translate3d(0px, ${-ITEM_HEIGHT * 2}px, 0px)`);
-
-      dispatchMouseEvent(document, 'mouseup');
-      fixture.detectChanges();
-      flush();
-    }));
-
-    it('should lay out the elements correctly, when swapping up with a taller element', fakeAsync(() => {
-      const fixture = createComponent(DraggableInDropZone);
-      fixture.detectChanges();
-
-      const items = fixture.componentInstance.dragItems.map(i => i.element.nativeElement);
-      const {top, left} = items[1].getBoundingClientRect();
-
-      fixture.componentInstance.items[1].height = ITEM_HEIGHT * 2;
-      fixture.changeDetectorRef.markForCheck();
-      fixture.detectChanges();
-
-      startDraggingViaMouse(fixture, items[1], left, top);
-
-      const placeholder = document.querySelector('.cdk-drag-placeholder')! as HTMLElement;
-      const target = items[0];
-      const targetRect = target.getBoundingClientRect();
-
-      // Add a few pixels to the top offset so we get some overlap.
-      dispatchMouseEvent(document, 'mousemove', targetRect.left, targetRect.bottom - 5);
-      fixture.detectChanges();
-
-      expect(placeholder.style.transform).toBe(`translate3d(0px, ${-ITEM_HEIGHT}px, 0px)`);
-      expect(target.style.transform).toBe(`translate3d(0px, ${ITEM_HEIGHT * 2}px, 0px)`);
-
-      dispatchMouseEvent(document, 'mouseup');
-      fixture.detectChanges();
-      flush();
-    }));
-
-    it('should lay out elements correctly, when swapping an item with margin', fakeAsync(() => {
-      const fixture = createComponent(DraggableInDropZone);
-      fixture.detectChanges();
-
-      const items = fixture.componentInstance.dragItems.map(i => i.element.nativeElement);
-      const {top, left} = items[0].getBoundingClientRect();
-
-      fixture.componentInstance.items[0].margin = 12;
-      fixture.changeDetectorRef.markForCheck();
-      fixture.detectChanges();
-
-      startDraggingViaMouse(fixture, items[0], left, top);
-
-      const placeholder = document.querySelector('.cdk-drag-placeholder')! as HTMLElement;
-      const target = items[1];
-      const targetRect = target.getBoundingClientRect();
-
-      // Add a few pixels to the top offset so we get some overlap.
-      dispatchMouseEvent(document, 'mousemove', targetRect.left, targetRect.top + 5);
-      fixture.detectChanges();
-
-      expect(placeholder.style.transform).toBe(`translate3d(0px, ${ITEM_HEIGHT + 12}px, 0px)`);
-      expect(target.style.transform).toBe(`translate3d(0px, ${-ITEM_HEIGHT - 12}px, 0px)`);
-
-      dispatchMouseEvent(document, 'mouseup');
-      fixture.detectChanges();
-      flush();
-    }));
-
-    it(
-      'should lay out the elements correctly, if an element skips multiple positions when ' +
-        'sorting horizontally',
-      fakeAsync(() => {
-        const fixture = createComponent(DraggableInHorizontalDropZone);
-        fixture.detectChanges();
-
-        const items = fixture.componentInstance.dragItems.map(i => i.element.nativeElement);
-        const draggedItem = items[0];
-        const {top, left} = draggedItem.getBoundingClientRect();
-
-        startDraggingViaMouse(fixture, draggedItem, left, top);
-
-        const placeholder = document.querySelector('.cdk-drag-placeholder')! as HTMLElement;
-        const targetRect = items[items.length - 1].getBoundingClientRect();
-
-        // Add a few pixels to the left offset so we get some overlap.
-        dispatchMouseEvent(document, 'mousemove', targetRect.right - 5, targetRect.top);
-        fixture.detectChanges();
-
-        expect(
-          config.getElementSibligsByPosition(placeholder, 'left').map(e => e.textContent!.trim()),
-        ).toEqual(['One', 'Two', 'Three', 'Zero']);
-
-        dispatchMouseEvent(document, 'mouseup');
-        fixture.detectChanges();
-        flush();
-      }),
-    );
-
-    it('should lay out the elements correctly, when swapping to the right with a wider element', fakeAsync(() => {
+    it('should lay out the elements correctly, if an element skips multiple positions when sorting horizontally', fakeAsync(() => {
       const fixture = createComponent(DraggableInHorizontalDropZone);
       fixture.detectChanges();
 
       const items = fixture.componentInstance.dragItems.map(i => i.element.nativeElement);
+      const draggedItem = items[0];
+      const {top, left} = draggedItem.getBoundingClientRect();
 
-      fixture.componentInstance.items[0].width = ITEM_WIDTH * 2;
-      fixture.changeDetectorRef.markForCheck();
-      fixture.detectChanges();
-
-      const {top, left} = items[0].getBoundingClientRect();
-      startDraggingViaMouse(fixture, items[0], left, top);
+      startDraggingViaMouse(fixture, draggedItem, left, top);
 
       const placeholder = document.querySelector('.cdk-drag-placeholder')! as HTMLElement;
-      const target = items[1];
-      const targetRect = target.getBoundingClientRect();
+      const targetRect = items[items.length - 1].getBoundingClientRect();
 
+      // Add a few pixels to the left offset so we get some overlap.
       dispatchMouseEvent(document, 'mousemove', targetRect.right - 5, targetRect.top);
       fixture.detectChanges();
 
-      expect(placeholder.style.transform).toBe(`translate3d(${ITEM_WIDTH}px, 0px, 0px)`);
-      expect(target.style.transform).toBe(`translate3d(${-ITEM_WIDTH * 2}px, 0px, 0px)`);
-
-      dispatchMouseEvent(document, 'mouseup');
-      fixture.detectChanges();
-      flush();
-    }));
-
-    it('should lay out the elements correctly, when swapping left with a wider element', fakeAsync(() => {
-      const fixture = createComponent(DraggableInHorizontalDropZone);
-      fixture.detectChanges();
-
-      const items = fixture.componentInstance.dragItems.map(i => i.element.nativeElement);
-      const {top, left} = items[1].getBoundingClientRect();
-
-      fixture.componentInstance.items[1].width = ITEM_WIDTH * 2;
-      fixture.changeDetectorRef.markForCheck();
-      fixture.detectChanges();
-
-      startDraggingViaMouse(fixture, items[1], left, top);
-
-      const placeholder = document.querySelector('.cdk-drag-placeholder')! as HTMLElement;
-      const target = items[0];
-      const targetRect = target.getBoundingClientRect();
-
-      dispatchMouseEvent(document, 'mousemove', targetRect.right - 5, targetRect.top);
-      fixture.detectChanges();
-
-      expect(placeholder.style.transform).toBe(`translate3d(${-ITEM_WIDTH}px, 0px, 0px)`);
-      expect(target.style.transform).toBe(`translate3d(${ITEM_WIDTH * 2}px, 0px, 0px)`);
-
-      dispatchMouseEvent(document, 'mouseup');
-      fixture.detectChanges();
-      flush();
-    }));
-
-    it('should lay out elements correctly, when horizontally swapping an item with margin', fakeAsync(() => {
-      const fixture = createComponent(DraggableInHorizontalDropZone);
-      fixture.detectChanges();
-
-      const items = fixture.componentInstance.dragItems.map(i => i.element.nativeElement);
-      const {top, left} = items[0].getBoundingClientRect();
-
-      fixture.componentInstance.items[0].margin = 12;
-      fixture.changeDetectorRef.markForCheck();
-      fixture.detectChanges();
-
-      startDraggingViaMouse(fixture, items[0], left, top);
-
-      const placeholder = document.querySelector('.cdk-drag-placeholder')! as HTMLElement;
-      const target = items[1];
-      const targetRect = target.getBoundingClientRect();
-
-      dispatchMouseEvent(document, 'mousemove', targetRect.right - 5, targetRect.top);
-      fixture.detectChanges();
-
-      expect(placeholder.style.transform).toBe(`translate3d(${ITEM_WIDTH + 12}px, 0px, 0px)`);
-      expect(target.style.transform).toBe(`translate3d(${-ITEM_WIDTH - 12}px, 0px, 0px)`);
+      expect(
+        config.getElementSibligsByPosition(placeholder, 'left').map(e => e.textContent!.trim()),
+      ).toEqual(['One', 'Two', 'Three', 'Zero']);
 
       dispatchMouseEvent(document, 'mouseup');
       fixture.detectChanges();
@@ -2065,30 +1897,26 @@ export function defineCommonDropListTests(config: {
       expect(Math.floor(previewRect.left)).toBe(50);
     }));
 
-    it(
-      'should revert the element back to its parent after dragging with a custom ' +
-        'preview has stopped',
-      fakeAsync(() => {
-        const fixture = createComponent(DraggableInDropZoneWithCustomPreview);
-        fixture.detectChanges();
+    it('should revert the element back to its parent after dragging with a custom preview has stopped', fakeAsync(() => {
+      const fixture = createComponent(DraggableInDropZoneWithCustomPreview);
+      fixture.detectChanges();
 
-        const dragContainer = fixture.componentInstance.dropInstance.element.nativeElement;
-        const item = fixture.componentInstance.dragItems.toArray()[1].element.nativeElement;
+      const dragContainer = fixture.componentInstance.dropInstance.element.nativeElement;
+      const item = fixture.componentInstance.dragItems.toArray()[1].element.nativeElement;
 
-        expect(dragContainer.contains(item))
-          .withContext('Expected item to be in container.')
-          .toBe(true);
+      expect(dragContainer.contains(item))
+        .withContext('Expected item to be in container.')
+        .toBe(true);
 
-        // The coordinates don't matter.
-        dragElementViaMouse(fixture, item, 10, 10);
-        flush();
-        fixture.detectChanges();
+      // The coordinates don't matter.
+      dragElementViaMouse(fixture, item, 10, 10);
+      flush();
+      fixture.detectChanges();
 
-        expect(dragContainer.contains(item))
-          .withContext('Expected item to be returned to container.')
-          .toBe(true);
-      }),
-    );
+      expect(dragContainer.contains(item))
+        .withContext('Expected item to be returned to container.')
+        .toBe(true);
+    }));
 
     it('should position custom previews next to the pointer', fakeAsync(() => {
       const fixture = createComponent(DraggableInDropZoneWithCustomPreview);
@@ -2213,22 +2041,18 @@ export function defineCommonDropListTests(config: {
       expect(preview.style.height).toBe(`${itemRect.height}px`);
     }));
 
-    it(
-      'should preserve the pickup position if the custom preview inherits the size of the ' +
-        'dragged element',
-      fakeAsync(() => {
-        const fixture = createComponent(DraggableInDropZoneWithCustomPreview);
-        fixture.componentInstance.matchPreviewSize = true;
-        fixture.detectChanges();
-        const item = fixture.componentInstance.dragItems.toArray()[1].element.nativeElement;
+    it('should preserve the pickup position if the custom preview inherits the size of the dragged element', fakeAsync(() => {
+      const fixture = createComponent(DraggableInDropZoneWithCustomPreview);
+      fixture.componentInstance.matchPreviewSize = true;
+      fixture.detectChanges();
+      const item = fixture.componentInstance.dragItems.toArray()[1].element.nativeElement;
 
-        startDraggingViaMouse(fixture, item, 50, 50);
+      startDraggingViaMouse(fixture, item, 50, 50);
 
-        const preview = document.querySelector('.cdk-drag-preview')! as HTMLElement;
+      const preview = document.querySelector('.cdk-drag-preview')! as HTMLElement;
 
-        expect(preview.style.transform).toBe('translate3d(8px, 33px, 0px)');
-      }),
-    );
+      expect(preview.style.transform).toBe('translate3d(8px, 33px, 0px)');
+    }));
 
     it('should not have the size of the inserted preview affect the size applied via matchSize', fakeAsync(() => {
       const fixture = createComponent(DraggableInHorizontalFlexDropZoneWithMatchSizePreview);
@@ -2362,30 +2186,6 @@ export function defineCommonDropListTests(config: {
 
       expect(placeholder).toBeTruthy();
       expect(placeholder.textContent!.trim()).toContain('HelloOne');
-    }));
-
-    it('should clear the `transform` value from siblings when item is dropped`', fakeAsync(() => {
-      const fixture = createComponent(DraggableInDropZone);
-      fixture.detectChanges();
-
-      const dragItems = fixture.componentInstance.dragItems;
-      const firstItem = dragItems.first;
-      const thirdItem = dragItems.toArray()[2].element.nativeElement;
-      const thirdItemRect = thirdItem.getBoundingClientRect();
-
-      startDraggingViaMouse(fixture, firstItem.element.nativeElement);
-
-      dispatchMouseEvent(document, 'mousemove', thirdItemRect.left + 1, thirdItemRect.top + 1);
-      fixture.detectChanges();
-
-      expect(thirdItem.style.transform).toBeTruthy();
-
-      dispatchMouseEvent(document, 'mouseup');
-      fixture.detectChanges();
-      flush();
-      fixture.detectChanges();
-
-      expect(thirdItem.style.transform).toBeFalsy();
     }));
 
     it('should not move the item if the list is disabled', fakeAsync(() => {
@@ -3358,57 +3158,6 @@ export function defineCommonDropListTests(config: {
         createComponent(DropListOnNgContainer).detectChanges();
         flush();
       }).toThrowError(/^cdkDropList must be attached to an element node/);
-    }));
-
-    it('should preserve the original `transform` of items in the list', fakeAsync(() => {
-      const fixture = createComponent(DraggableInScrollableVerticalDropZone);
-      fixture.detectChanges();
-      const items = fixture.componentInstance.dragItems.map(item => item.element.nativeElement);
-      items.forEach(element => (element.style.transform = 'rotate(180deg)'));
-      const thirdItemRect = items[2].getBoundingClientRect();
-      const hasInitialTransform = (element: HTMLElement) =>
-        element.style.transform.indexOf('rotate(180deg)') > -1;
-
-      startDraggingViaMouse(fixture, items[0]);
-      fixture.detectChanges();
-      const preview = document.querySelector('.cdk-drag-preview') as HTMLElement;
-      const placeholder = fixture.nativeElement.querySelector('.cdk-drag-placeholder');
-
-      expect(items.every(hasInitialTransform))
-        .withContext('Expected items to preserve transform when dragging starts.')
-        .toBe(true);
-      expect(hasInitialTransform(preview))
-        .withContext('Expected preview to preserve transform when dragging starts.')
-        .toBe(true);
-      expect(hasInitialTransform(placeholder))
-        .withContext('Expected placeholder to preserve transform when dragging starts.')
-        .toBe(true);
-
-      dispatchMouseEvent(document, 'mousemove', thirdItemRect.left + 1, thirdItemRect.top + 1);
-      fixture.detectChanges();
-      expect(items.every(hasInitialTransform))
-        .withContext('Expected items to preserve transform while dragging.')
-        .toBe(true);
-      expect(hasInitialTransform(preview))
-        .withContext('Expected preview to preserve transform while dragging.')
-        .toBe(true);
-      expect(hasInitialTransform(placeholder))
-        .withContext('Expected placeholder to preserve transform while dragging.')
-        .toBe(true);
-
-      dispatchMouseEvent(document, 'mouseup');
-      fixture.detectChanges();
-      flush();
-      fixture.detectChanges();
-      expect(items.every(hasInitialTransform))
-        .withContext('Expected items to preserve transform when dragging stops.')
-        .toBe(true);
-      expect(hasInitialTransform(preview))
-        .withContext('Expected preview to preserve transform when dragging stops.')
-        .toBe(true);
-      expect(hasInitialTransform(placeholder))
-        .withContext('Expected placeholder to preserve transform when dragging stops.')
-        .toBe(true);
     }));
 
     it('should sort correctly if the <html> node has been offset', fakeAsync(() => {
@@ -5037,9 +4786,15 @@ export function defineCommonDropListTests(config: {
       expect(event.stopPropagation).toHaveBeenCalled();
     }));
   });
+}
 
-  // Horizontal fixtures need to be defined here, because the
-  // `horizontalListOrientation` value can change between tests.
+/**
+ * Dynamically creates the horizontal list fixtures. They need to be
+ * generated so that the list orientation can be changed between tests.
+ * @param listOrientation Orientation value to be assigned to the list.
+ *    Does not affect the actual styles.
+ */
+export function getHorizontalFixtures(listOrientation: Exclude<DropListOrientation, 'vertical'>) {
   // Use inline blocks here to avoid flexbox issues and not to have to flip floats in rtl.
   const HORIZONTAL_FIXTURE_STYLES = `
     .cdk-drop-list {
@@ -5060,7 +4815,7 @@ export function defineCommonDropListTests(config: {
     <div
       class="drop-list scroll-container"
       cdkDropList
-      cdkDropListOrientation="${config.horizontalListOrientation}"
+      cdkDropListOrientation="${listOrientation}"
       [cdkDropListData]="items"
       (cdkDropListDropped)="droppedSpy($event)">
       @for (item of items; track item) {
@@ -5092,7 +4847,7 @@ export function defineCommonDropListTests(config: {
       moveItemInArray(this.items, event.previousIndex, event.currentIndex);
     });
 
-    constructor(protected _elementRef: ElementRef) {}
+    constructor(readonly _elementRef: ElementRef) {}
 
     ngAfterViewInit() {
       // Firefox preserves the `scrollLeft` value from previous similar containers. This
@@ -5145,7 +4900,7 @@ export function defineCommonDropListTests(config: {
       }
     `,
     template: `
-      <div class="list" cdkDropList cdkDropListOrientation="${config.horizontalListOrientation}">
+      <div class="list" cdkDropList cdkDropListOrientation="${listOrientation}">
         @for (item of items; track item) {
           <div class="item" cdkDrag>
             {{item}}
@@ -5161,6 +4916,12 @@ export function defineCommonDropListTests(config: {
     @ViewChildren(CdkDrag) dragItems: QueryList<CdkDrag>;
     items = ['Zero', 'One', 'Two'];
   }
+
+  return {
+    DraggableInHorizontalDropZone,
+    DraggableInScrollableHorizontalDropZone,
+    DraggableInHorizontalFlexDropZoneWithMatchSizePreview,
+  };
 }
 
 // TODO(crisbeto): figure out why switch `*ngFor` with `@for` here causes a test failure.
@@ -5191,8 +4952,12 @@ const DROP_ZONE_FIXTURE_TEMPLATE = `
   <div #alternatePreviewContainer></div>
 `;
 
-@Component({template: DROP_ZONE_FIXTURE_TEMPLATE})
-class DraggableInDropZone implements AfterViewInit {
+@Component({
+  template: DROP_ZONE_FIXTURE_TEMPLATE,
+  standalone: true,
+  imports: [CdkDropList, CdkDrag, NgFor],
+})
+export class DraggableInDropZone implements AfterViewInit {
   @ViewChildren(CdkDrag) dragItems: QueryList<CdkDrag>;
   @ViewChild(CdkDropList) dropInstance: CdkDropList;
   @ViewChild('alternatePreviewContainer') alternatePreviewContainer: ElementRef<HTMLElement>;
@@ -5254,7 +5019,7 @@ class ConnectedDropListsInOnPush {}
     }
   `,
 })
-class DraggableInScrollableVerticalDropZone extends DraggableInDropZone {
+export class DraggableInScrollableVerticalDropZone extends DraggableInDropZone {
   constructor(elementRef: ElementRef) {
     super(elementRef);
