+++ src/cdk/drag-drop/directives/test-utils.spec.ts
@@ -1,4 +1,10 @@
-import {EnvironmentProviders, Provider, Type, ViewEncapsulation} from '@angular/core';
+import {
+  EnvironmentProviders,
+  Provider,
+  Type,
+  ViewEncapsulation,
+  reflectComponentType,
+} from '@angular/core';
 import {ComponentFixture, TestBed, tick} from '@angular/core/testing';
 import {dispatchMouseEvent, dispatchTouchEvent} from '@angular/cdk/testing/private';
 import {CdkScrollableModule} from '@angular/cdk/scrolling';
@@ -23,6 +29,8 @@ export function createComponent<T>(
   componentType: Type<T>,
   config: DragDropTestConfig = {},
 ): ComponentFixture<T> {
+  // TODO(crisbeto): drop this logic once all the fixtures are converted to standalone.
+  const isStandalone = reflectComponentType(componentType)?.isStandalone;
   const dragConfig: DragDropConfig = {
     // We default the `dragDistance` to zero, because the majority of the tests
     // don't care about it and drags are a lot easier to simulate when we don't
@@ -31,6 +39,11 @@ export function createComponent<T>(
     pointerDirectionChangeThreshold: 5,
     listOrientation: config.listOrientation,
   };
+  const declarations = [...(config.extraDeclarations || [])];
+
+  if (!isStandalone) {
+    declarations.push(componentType);
+  }
 
   TestBed.configureTestingModule({
     imports: [DragDropModule, CdkScrollableModule],
@@ -41,7 +54,7 @@ export function createComponent<T>(
       },
       ...(config.providers || []),
     ],
-    declarations: [componentType, ...(config.extraDeclarations || [])],
+    declarations,
   });
 
   if (config.encapsulation != null) {
